requirements.txt
# MomaCleaner Requirements
pyyaml>=6.0
Pillow>=10.0.0
